api: rename _env.py to env.py
The leading underscore signals "private/internal" in Python, but this
module is a generic helper that other api/ modules import. That matches
the convention for cache.py, types.py, scan.py, etc. — not the dev-only
_reload.py outlier. Renaming for consistency.
